#44 refactor: select 페이지 Suspense 경계 도입으로 선언적 로딩 처리
