+++ app/letter/new/select/page.tsx
@@ -1,80 +1,22 @@
 "use client";
 
-import { useState } from "react";
-import { useGetLetterData } from "@/hooks/apis/get/useGetLetterData";
-import { useAudioPlayer } from "@/hooks/common/useAudioPlayer";
-import CompleteButtonContainer from "@/components/select/CompleteButtonContainer";
-import LetterBox from "@/components/select/LetterBox";
-import AudioPlayer from "@/components/select/AudioPlayer";
-import LetterOptionsBottomSheet from "@/components/select/LetterOptionsBottomSheet";
-import FloatingButton from "@/components/select/FloatingButton";
-
-type TabType = "font" | "paper" | "bgm";
+import { Suspense } from "react";
+import { Spinner } from "@/components/ui/spinner";
+import SelectPageContent from "@/components/select/SelectPageContent";
 
 export default function SelectPage() {
-  const { data } = useGetLetterData();
-  const [isSheetOpen, setIsSheetOpen] = useState(false);
-  const [selectedTab, setSelectedTab] = useState<TabType>("font");
-
-  const {
-    status,
-    currentTime,
-    progress,
-    duration,
-    togglePlayPause,
-    stop,
-    seek,
-  } = useAudioPlayer({
-    audioUrl: data?.voiceUrl || null,
-    initialDuration: data?.duration || 0,
-  });
-
-  if (!data) return null;
-
   return (
-    <article className="bg-gray-50 h-full relative">
-      <section className="flex flex-col p-5 gap-8 h-full justify-between">
-        <div className="flex flex-col gap-5 justify-between h-full">
-          <LetterBox
-            title={data.title}
-            sender={data.sender}
-            content={data.content}
-            fontId={data.fontId}
-            fontUrl={data.fontUrl}
-            templateUrl={data.templateUrl}
-            isEdit={true}
-            words={data.words}
-            currentTime={currentTime}
-            status={status}
-          />
-          <AudioPlayer
-            bgmUrl={data.bgmUrl}
-            bgmSize={data.bgmSize}
-            status={status}
-            currentTime={currentTime}
-            progress={progress}
-            duration={duration}
-            togglePlayPause={togglePlayPause}
-            stop={stop}
-            seek={seek}
-          />
+    <Suspense
+      fallback={
+        <div className="flex flex-col items-center justify-center h-[80%] gap-4">
+          <p className="text-gray-400 text-center typo-h2-lg">
+            편지 불러오는 중...
+          </p>
+          <Spinner className="size-12 text-gray-400" />
         </div>
-
-        <CompleteButtonContainer />
-      </section>
-
-      <FloatingButton
-        onOptionClick={(option) => {
-          setSelectedTab(option);
-          setIsSheetOpen(true);
-        }}
-      />
-
-      <LetterOptionsBottomSheet
-        isOpen={isSheetOpen}
-        onOpenChange={setIsSheetOpen}
-        initialTab={selectedTab}
-      />
-    </article>
+      }
+    >
+      <SelectPageContent />
+    </Suspense>
   );
 }
